condenses routes
