+++ lib/auth.js
@@ -22,6 +22,7 @@ module.exports = function(log, conf) {
         }
 
         var user = basicAuth(req);
+        console.log(req.body);
         if (checkPassword(user && user.pass || '')) {
             return next();
         } else {
