+++ routes/granary.js
@@ -4,6 +4,9 @@ var path = require('path');
 var kue = require('kue');
 var Job = require('kue/lib/queue/job');
 var reds = require('reds');
+var filesize = require('filesize');
+var async = require('async');
+var moment = require('moment');
 
 module.exports = function(log, conf) {
 
@@ -27,9 +30,63 @@ module.exports = function(log, conf) {
     var freighter = require('../lib/freighter')(log, conf, jobs);
     var tracker = require('../lib/tracker')(log, conf, jobs);
     var freightAuth = require('../lib/auth')(log, conf);
-    var FreightRoutes = {};
+    var Routes = {};
 
-    FreightRoutes.check = function(req, res) {
+    Routes.usage = function(req, res, next) {
+        var memory = process.memoryUsage();
+        var storage = conf.get('storage');
+
+        var data = {
+            title: 'Granary Server',
+            process: {
+                heap: filesize(memory.heapUsed)
+            }
+        };
+
+        fs.readdir(storage, function(err, files) {
+            if (err) {
+                log.error(err);
+                throw err;
+            }
+
+            async.map(files,
+                function(file, complete) {
+                    fs.stat(path.join(storage, file), function(err, stat) {
+                        stat.size = filesize(stat.size);
+                        stat.download = '/storage/' + file;
+                        stat.ctime = moment(stat.ctime).format('MMMM Do YYYY, h:mm:ss a');
+                        if (file.indexOf('tar.gz') > -1) {
+                            complete(err, {
+                                name: file,
+                                details: stat
+                            });
+                        } else {
+                            complete(err, {});
+                        }
+                    });
+                },
+                function(err, results) {
+                    if (err) {
+                        log.error('Failed to get statistics on all bundle files');
+                        throw err;
+                    }
+
+                    var bundles = results.filter(function(file) {
+                        return file.hasOwnProperty('name');
+                    });
+
+                    bundles.sort(function(a, b) {
+                        return b.details.mtime.getTime() - a.details.mtime.getTime();
+                    });
+
+                    data.files = bundles;
+                    req.data = data;
+                    next();
+                });
+        });
+    }
+
+    Routes.check = function(req, res) {
         if (!req.body && !req.body.project && !req.body.project.name) {
             return res.sendStatus(404);
         }
@@ -97,7 +154,7 @@ module.exports = function(log, conf) {
         });
     };
 
-    FreightRoutes.download = function(req, res) {
+    Routes.download = function(req, res) {
         log.debug('Download request', req.body);
         if (req.body.hash) {
             var hashFile = path.join(conf.get('storage'), req.body.name + '-' + req.body.hash + '.tar.gz');
@@ -121,7 +178,7 @@ module.exports = function(log, conf) {
 
     };
 
-    FreightRoutes.track = function(req, res) {
+    Routes.track = function(req, res) {
         if (req.body && req.body.repository && req.body.password && req.body.branch) {
             log.debug('Tracking request:', req.body);
 
@@ -153,5 +210,5 @@ module.exports = function(log, conf) {
 
     };
 
-    return FreightRoutes;
+    return Routes;
 };
